# internal/apidocs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Expert Amp Server API",
    "version": "0.1.0",
    "description": "Trusted-LAN HTTP API for Expert Amp Server. The service is intended for local station networks and should not be exposed directly to the public internet. The canonical phase 1 surface lives under /api/v1. This spec intentionally documents current implementation reality, including conservative or fallback-derived fields. It does not treat legacy compatibility routes as canonical, and it does not list older non-v1 aliases except where needed for explanatory context.\n\nStatus honesty rules:\n- GET /api/v1/status is the preferred machine-readable amp status surface.\n- Status fields may be protocol-native, display-derived fallback, compatibility aliases, or current conservative placeholders depending on what the runtime actually knows.\n- GET /api/v1/telemetry remains useful for UI and fallback inspection, but it is not the authoritative automation contract when protocol-native status poll data is available.\n- GET /api/v1/alarms is still a phase 1 stub with an explicit stub flag.\n- The websocket status feed shares the same canonical status model but is described separately because OpenAPI does not model websocket message flow well."
  },
  "servers": [
    {
      "url": "http://localhost:8088",
      "description": "Typical local development or LAN deployment"
    }
  ],
...
  "paths": {
...
    "/api/v1/display/state": {
      "get": {
        "tags": [
          "Display"
        ],
        "summary": "Get display state",
        "description": "Returns the current 8x40 runtime display state by default. Optional fixture selection is for protocol inspection and does not mutate runtime state.",
        "parameters": [
          {
            "$ref": "#/components/parameters/KindQuery"
          },
          {
            "$ref": "#/components/parameters/SourceQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/DisplayStateResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "405": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/display/frame": {
      "get": {
        "tags": [
          "Display"
        ],
        "summary": "Get frame metadata",
        "description": "Returns runtime frame metadata by default. Optional fixture selection returns fixture frame metadata instead.",
        "parameters": [
          {
            "$ref": "#/components/parameters/KindQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/FrameInfoResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "405": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/display/render.png": {
      "get": {
        "tags": [
          "Display"
        ],
        "summary": "Render display as PNG",
        "parameters": [
          {
            "$ref": "#/components/parameters/KindQuery"
          },
          {
            "$ref": "#/components/parameters/SourceQuery"
          },
          {
            "$ref": "#/components/parameters/ScaleQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "PNG render of the selected display state",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "405": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/display/render.svg": {
      "get": {
        "tags": [
          "Display"
        ],
        "summary": "Render display as SVG",
        "parameters": [
          {
            "$ref": "#/components/parameters/KindQuery"
          },
          {
            "$ref": "#/components/parameters/SourceQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "SVG render of the selected display state",
            "content": {
              "image/svg+xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "405": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/display/ws": {
      "get": {
        "tags": [
          "Display"
        ],
        "summary": "Display refresh websocket feed",
        "description": "WebSocket endpoint that emits a lightweight event each time the authoritative runtime display snapshot changes. The watch UI uses this to refresh /api/v1/display/render.png when a new frame is available instead of blind browser polling. Legacy HTTP render endpoints remain available for polling consumers.",
        "responses": {
          "101": {
            "description": "WebSocket upgrade accepted. Text messages carry display refresh events with sequence, updatedAt, source, and frameKind."
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "405": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/status": {
      "get": {
        "tags": [
          "Status"
        ],
        "summary": "Get canonical amp status",
        "description": "Preferred automation surface. Protocol-native status poll data wins when present. Display-derived telemetry fills gaps conservatively. Some fields remain omitted rather than guessed.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusResponse"
          },
          "405": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/status/ws": {
      "get": {
        "tags": [
          "Status"
        ],
        "summary": "Status websocket feed",
        "description": "WebSocket endpoint for the same canonical status model returned by GET /api/v1/status. After connect, the server sends one status snapshot immediately and then pushes changed snapshots only. Delivery is event-driven from the authoritative status path. The canonical endpoint does not support pace query strings. OpenAPI does not model websocket frames cleanly, so this operation is documented descriptively rather than with a full message schema exchange.",
        "responses": {
          "101": {
            "description": "WebSocket upgrade accepted. Text messages carry Status objects without the standard envelope."
          },
          "405": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/telemetry": {
      "get": {
        "tags": [
          "Status"
        ],
        "summary": "Get runtime telemetry snapshot",
        "description": "UI-friendly runtime telemetry and fallback inspection surface. Useful, but not the canonical machine status contract when protocol-native status exists.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/TelemetryResponse"
          },
          "405": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/alarms": {
      "get": {
        "tags": [
          "Status"
        ],
        "summary": "Get phase 1 alarms view",
        "description": "Current implementation returns an honest stub with active: [] and stub: true.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/AlarmsResponse"
          },
          "405": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/runtime": {
      "get": {
        "tags": [
          "Runtime"
        ],
        "summary": "Get effective runtime polling view",
        "responses": {
          "200": {
            "$ref": "#/components/responses/RuntimeStatusResponse"
          },
          "405": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/runtime/snapshot": {
      "get": {
        "tags": [
          "Runtime"
        ],
        "summary": "Get full runtime snapshot",
        "responses": {
          "200": {
            "$ref": "#/components/responses/RuntimeSnapshotResponse"
          },
          "405": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/runtime/ingest": {
      "get": {
        "tags": [
          "Runtime"
        ],
        "summary": "Get serial ingest diagnostics",
        "description": "In fixture mode this returns connected=false and a fixture-mode explanation instead of pretending a live serial path exists.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/IngestDiagnosticsResponse"
          },
          "405": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/runtime/restart": {
      "post": {
        "tags": [
          "Runtime"
        ],
        "summary": "Request a clean process restart",
        "description": "Requests a clean process exit. A host supervisor such as systemd, launchd, or a Docker restart policy must bring the process back. If no supervisor is present, the process stops and stays stopped.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/BasicSuccessResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "405": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get current config snapshot",
        "responses": {
          "200": {
            "$ref": "#/components/responses/SettingsResponse"
          },
          "405": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Update settings",
        "description": "Writes local config. displayPollingEnabled and statusPollingEnabled are live toggles. Listen address and pollIntervalMs (the backend display-frame refresh cadence) still require restart to take effect. statusPollIntervalMs controls backend protocol-status refresh. Legacy serialPoll* fields remain accepted only as backward-compatible aliases; the documented status poll frame itself is fixed in code rather than user-editable.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdateRequest"
              },
              "examples": {
                "basic": {
                  "value": {
                    "serialPort": "/dev/ttyUSB0",
                    "listenAddress": ":8088",
                    "pollIntervalMs": 200,
                    "displayPollingEnabled": true,
                    "statusPollingEnabled": true,
                    "serialBaudRate": 115200,
                    "serialReadTimeoutMs": 250,
                    "statusPollCommandEnabled": true,
                    "statusPollIntervalMs": 125,
                    "serialAssertDTR": true,
                    "serialAssertRTS": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SettingsResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "405": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/serial-ports": {
      "get": {
        "tags": [
          "Runtime"
        ],
        "summary": "Enumerate visible serial ports",
        "responses": {
          "200": {
            "$ref": "#/components/responses/SerialPortsResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "405": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "KindQuery": {
        "name": "kind",
        "in": "query",
        "description": "Optional fixture kind for protocol inspection. Supported values depend on built-in fixtures, typically home, menu, or panel.",
        "schema": {
          "type": "string"
        }
      },
      "SourceQuery": {
        "name": "source",
        "in": "query",
        "description": "Optional fixture source selector. source=protocol with no explicit kind currently defaults to the home fixture.",
        "schema": {
          "type": "string",
          "enum": [
            "protocol"
          ]
        }
      },
...
    },
    "responses": {
...
      "DisplayStateResponse": {
        "description": "Display state envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ResponseDisplayState"
            }
          }
        }
      },
...
    },
...
  }
}